refactored /workout, created Exercise component
