+++ apps/gym/src/app/record/workout/Workout.module.css
@@ -36,91 +36,3 @@
   padding-bottom: 10px;
 }
 
-.single_exercise {
-  margin: 5px;
-  padding: 20px;
-}
-
-.exercise_name {
-  display: inline-block;
-  font-size: 1.3rem;
-  font-weight: 550;
-  color: var(--blue);
-  padding: 5px;
-  border-radius: 4px;
-}
-
-.exercise_header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0;
-}
-
-.single_exercise hr {
-  margin: 0px;
-  border-top: 2px solid var(--blue);
-  margin-bottom: 10px;
-  opacity: 35%;
-}
-
-.set_labels label {
-  display: inline-block;
-  width: 59px;
-  margin: 0 15px;
-  font-size: 1.1rem;
-}
-
-#set_label {
-  margin: 0;
-}
-
-.notes {
-  display: flex;
-  flex-direction: column;
-  margin: 10px 0 10px 0;
-}
-
-.notes textarea {
-  border: 2px solid lightgray;
-  border-radius: 4px;
-  padding: 6px 4px;
-  height: 18px;
-  outline: none;
-  width: 100%;
-  resize: none;
-  max-width: 370px;
-  font-size: 1.15rem;
-  font-family: inherit;
-  line-height: normal;
-}
-
-.notes textarea:is(:focus) {
-  border-color: var(--blue);
-}
-
-.notes textarea::-webkit-scrollbar {
-  display: none;
-}
-
-.add_set {
-  display: block;
-  margin: 10px auto;
-  border: none;
-  font-size: 1.1rem;
-  padding: 3px 60px;
-  background-color: rgba(58, 143, 234, .8);
-  border-radius: 8px;
-  line-height: 24px;
-  font-family: inherit;
-}
-
-.hamburger_container {
-  width: 100%;
-  height: 100%;
-  position: relative;
-}
-
-.three_dots {
-  cursor: pointer;
-}
