+++ apps/gym/src/app/record/workout/finished/Finished.module.css
@@ -8,15 +8,4 @@
   align-items: center;
   justify-content: center;
   font-size: 2rem;
-  z-index: -1;
-}
-
-.home_link {
-  color: black;
-  text-decoration: none;
-  padding: 20px 40px;
-  border-radius: 40px;
-  background-color: var(--blue);
-  margin-top: 20px;
-  line-height: 48px;
 }
