+++ apps/gym/src/components/isolated/Record/Exercise/index.ts
@@ -1,0 +1,1 @@
+export * from './Exercise'
