+++ apps/gym/src/components/isolated/Record/index.ts
@@ -2,3 +2,4 @@ export * from './AddExerciseModal'
 export * from './ExerciseModal'
 export * from './Set'
 export * from './WorkoutConfirmationModal'
+export * from './Exercise'
