+++ apps/gym/src/components/reusable/Navbar/Navbar.module.css
@@ -6,6 +6,8 @@
     max-height: 60px;
     justify-content: space-between;
     align-items: center;
+    position: relative;
+    z-index: 10000;
 }
 
 .logo {
